+++ optimum/neuron/accelerate/utils/misc.py
@@ -16,20 +16,19 @@
 
 import functools
 import inspect
-from typing import TYPE_CHECKING, Dict, Optional, Union
+from typing import TYPE_CHECKING, Union
 
 import torch
 
 from ....utils import logging
 from ...utils import is_torch_neuronx_available, is_torch_xla_available, patch_everywhere
 from ...utils.peft_utils import NeuronPeftModel
-from ...utils.require_utils import requires_neuronx_distributed, requires_safetensors, requires_torch_xla
+from ...utils.require_utils import requires_neuronx_distributed, requires_torch_xla
 
 
 logger = logging.get_logger(__name__)
 
 if TYPE_CHECKING:
-    import os
 
     from transformers import PreTrainedModel
 
@@ -58,28 +57,6 @@ def patch_accelerate_is_torch_xla_available():
     patch_everywhere(
         "is_torch_xla_available", patched_accelerate_is_torch_xla_available, module_name_prefix="accelerate"
     )
-
-
-@requires_neuronx_distributed
-@requires_safetensors
-def torch_xla_safe_save_file(
-    tensors: Dict[str, torch.Tensor],
-    filename: Union[str, "os.PathLike"],
-    metadata: Optional[Dict[str, str]] = None,
-    master_only: bool = True,
-    global_master: bool = False,
-):
-    """
-    Torch XLA compatible implementation of `safetensors.torch.save_file`.
-    """
-    from neuronx_distributed.parallel_layers.utils import move_all_tensor_to_cpu
-    from safetensors.torch import save_file
-    from torch_xla.core.xla_model import is_master_ordinal
-
-    should_write_data = not master_only or is_master_ordinal(local=not global_master)
-    cpu_data = move_all_tensor_to_cpu(tensors, convert=should_write_data)
-    if should_write_data:
-        save_file(cpu_data, filename, metadata=metadata)
 
 
 # TODO: @michaelbenayoun
